[KYUUBI #360] move ThriftUtils to kyuubi-common
